requirements.txt
duckdb==1.5.0
numpy==2.4.3
pandas==3.0.1
psycopg2-binary==2.9.11
python-dateutil==2.9.0.post0
six==1.17.0
SQLAlchemy==2.0.48
tabulate==0.10.0
typing_extensions==4.15.0
